Merge uat to main

+++ orders/models.py
@@ -30,7 +30,7 @@ class OrderItem(models.Model):
 
     order = models.ForeignKey(Order, on_delete=models.CASCADE, related_name="items")
     prescription = models.ForeignKey(Prescription, on_delete=models.CASCADE, null=True, blank=True)
-    quantity = models.PositiveBigIntegerField(default=1)
+    quantity = models.PositiveIntegerField(default=1)
     price = models.DecimalField(max_digits=10, decimal_places=2)
 
     def __str__(self):
